sanlock: fix read_resource_owners read error
It would try to read an 8MB lease, but if the lease
was 1MB at the end of the file/device, then the read
would return an error because of the short read.
We can ignore this error.

Also, when we have to guess at the lease size to start
with, guess 1M instead of 8M.  If either guess is wrong
the read will be retried with the correct size.