python: Clean up alignment and sector constants
The constants used SANLK_RES_ prefix, which is not needed, confusing,
and make client code uglier.

- SANLK_ prefix is not needed since all the constants are in the sanlock
  module namespace (e.g sanlock.ALIGN1M).

- RES_ prefix is incorrect because the alignment and sector size are
  used for lockspace, rindex and resources.

The help for the python functions mentioned the C API constants names
(e.g. SANLK_LSF_ALIGN1M). Now we mention the python constant name (e.g
ALIGN1M) which is the value the user of this module must use.

Since this version was not released yet, we still have time to fix this
API.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>