python: Unify python extension make command
Replace the python Makefile inplace target with environment variable,
fixing the tests after the top Makefile was simplified to use:

$(SUBDIRS):
	$(MAKE) -C $@

When running tox, we build sanlock extension inside the tox relevant tox
environment (e.g. .tox/py27/...). This will be helpful for supporting
python 3.

On Travis we still build in the python directory - sanlock import fails
if we build it elsewhere. This is not an issue for python 3 since on
Travis we have separate build for every python version.

Fixes: 6133e3471ca3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>