Accept values instead of flags for sector and aling parameters.
Currently appropriate sanlock flags are accepted for sector size and
alignment parameters. This can be confusing for users, which may try
to pass values instead of flags and this approach also exposes sanlock
internal implementation details.

In order to have consistent and more intuitive API, values should be
accepted for sector size and alignment parameters. This patch replaces
flags by values in calls which accept sector size and alignment.

This patch also adds validation of provided values. If the value is
not the supported one, error with error code EINVAL is thrown.

The supported values are:
sector: 512, 4096
align: 1048576, 2097152, 4194304, 8388608

The patch also adds sector and align parameters to read and write
lockspace and resource calls to test_write_lockspace and
test_write_resource tests to test that these parameters are accepted.
Other tests still use calls without these parameters to test the
defaults. Tests for invalid align and sector values are also added.