python: Apply pypath_converter(...) for file paths
In Python 3 we would like to handle paths as either unicode strings
or bytes. For this purpose we have a c-API called PyUnicode_FSConverter.
In this patch the missing API is implemented for Py2 based on Py3's [1]
and a unified no-cleanup behavior is enforced in Py3 to comply with Py2's usage.

[1] https://github.com/python/cpython/blob/master/Objects/unicodeobject.c#L3818

Signed-off-by: Amit Bawer <abawer@redhat.com>