tests: Simplify 4k tests setup and teardown
Previously you had to setup the environment manually to enable the 4k
tests, and pass environment variable to tests. The tests used single
path, so you had to run the tests twice to test both 4k block device and
filesystem backed by a 4k block device.

Add storage helper script to make it easier to setup storage for the
tests.

To setup 4k storage for the tests, run:

   $ python tests/storage.py setup

This script creates both a 4k block device and filesystem backed by a 4k
device. The user_4k_path() fixture enables the tests if the device and
mount exist, and runs the test twice, once with the loop device, and
once with the filesystem.

To clean up the environment, run:

   $ python tests/storage.py teardown

Signed-off-by: Nir Soffer <nsoffer@redhat.com>