tests: Remove stale symlinks before linking loop device
On storage setup we can have stale symlinks to loop devices
which were created before reboot and now are gone. Re-running the
setup will fail on creating the symlink again. In this patch we
add removal of the pre-existing symlink, if there is such, before
symlinking the new loop device.