python: Add pypath_converter
In Python 3 we would like to handle paths as either unicode strings
or bytes. For this purpose we have a c-API called PyUnicode_FSConverter.
In this patch the missing API is implemented for Py2 based on Py3 [1]
and a unified no-cleanup behavior is enforced for Py3 path conversion as well.

[1] https://github.com/python/cpython/blob/master/Objects/unicodeobject.c#L3818