python: Replace async with wait
The async argument to add_lockspace() and rem_lockspace() is not valid
syntax in python 3.7, since async is a keyword now.

Replace async=False with wait=True, already used in other sanlock
methods.

This change breaks compatibility with existing python 2 code, but we
plan to ship sanlock 3.8 with python 3 support only on Fedora 30 and
EL 8.1 which do not support python 2, so this does not break existing
users.

Python code that wants to support both sanlock 3.7 and 3.8 will have to
change the behaviour based on python version.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>