sanlock: fix direct dump with offset
For direct dump <path>:<offset>:<size> the dump is meant
to continue reading/reporting for the entire range.  If
the sector_size/align_size were not at <offset> the dump
would stop.  Fix this by searching through the range for
sector_size/align_size before beginning the real dump.

Accept the M (MB) suffix on offset and size.

Fix the printed offset of structures so that it is always
relative to the start of the disk (it had been relative to
the <offset>.)