sanlock: write_init_io_timeout
Add a new io timeout used when initializing ondisk structures for
a lockspace or resource.  This is independent of the standard
io timeout used as part of the lockspace or resource algorithm.
This write_init_io_timeout will generally be larger, can be set
in sanlock.conf, and if zero the code will fall back to using
the standard io_timeout as has been used before.