userstorage: replace by PyPI userstorage
Remove the outdated version of userstorage
that is stored locally at tests/storage.py

The userstorage library is required
to create the storage before running the
tests with tox, and also in the test
virtual environment. In order to satisfy
this, add a requirements.txt to the root
folder to obtain the latests userstorage
version from PyPI, and add it as a tox
dependency for tests.

Behaviour for missing storage and folder
structure remains as originally intended.

Fixture is adapted to new style so that
is uses setup and teardown before and
after every test, respectively.

Signed-off-by: Albert Esteve <aesteve@redhat.com>